Persisted store, small redirect fix and small folder restructuring
