+++ client/package.json
@@ -51,7 +51,6 @@
     ]
   },
   "devDependencies": {
-    "@react-native-community/async-storage": "^1.12.1",
     "@testing-library/jest-dom": "^5.11.10",
     "@testing-library/react": "^11.2.5",
     "@testing-library/user-event": "^12.8.3",
