+++ client/src/ui/make-purchase-components/MakePurchaseComponent.tsx
@@ -5,6 +5,8 @@ import { GuestPurchaseComponent } from "./guest-purchase-component/GuestPurchase
 import { NotVerifiedPurchaseComponent } from "./not-verified-purchase-component/NotVerifiedPurchaseComponent";
 import { OkPurchaseComponent } from "./ok-purchase-component/OkPurchaseComponent";
 import { EmptyCartPurchaseComponent } from "./empty-cart-purchase-component/EmptyCartPurchaseComponent";
+import { logger } from "../../utils/logger";
+import storeAndPersistor from "../../store/store";
 
 export const MakePurchaseComponent: React.FC<MakePurchaseComponentProps> = (props) => {
   const { isVerified, productsInCart, emailAndPassword } = props;
@@ -19,5 +21,8 @@ export const MakePurchaseComponent: React.FC<MakePurchaseComponentProps> = (prop
   if (productsInCart.size === 0) {
     return <EmptyCartPurchaseComponent />;
   }
+  const { store } = storeAndPersistor();
+  const { currentUser, orderList, productList, shoppingCart } = store.getState();
+  logger.log(currentUser, orderList, productList, shoppingCart);
   return <OkPurchaseComponent productsInCart={productsInCart} />;
 };
